Test STIR2PETSIRD conversion and CI workflows.
